requirements.txt
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.7
cookiecutter==2.6.0
cryptography==44.0.2
defusedxml==0.7.1
distro==1.9.0
Django==5.1
django-allauth==65.5.0
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==25.1
django-htmx==1.22.0
django-jazzmin==3.0.1
django-meta==2.4.2
django-simple-history==3.7.0
django-snakeoil==2.0
django-tailwind==3.8.0
django-tenants==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
et-xmlfile==1.1.0
Flask==3.0.3
fpdf==1.7.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
meta==1.0.2
MouseInfo==0.1.3
numpy==2.1.1
odf==0.0.1
odfpy==1.4.1
openai==1.63.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pillow==10.4.0
psycopg2-binary==2.9.9
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.9.0
PyMsgBox==1.0.9
PyPDF2==3.0.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-pptx==1.0.2
python-slugify==8.0.4
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
text-unidecode==1.3
tqdm==4.67.1
typer==0.12.5
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.6
whitenoise==6.9.0
wikipedia==1.4.0
wkhtmltopdf==0.2
XlsxWriter==3.2.0
